Added msgpack2json continuous mode
Continuous mode allows converting a stream of top-level objects to JSON.
Comma delimiting is optional.

This includes a fix for a potential bug when printing the final newline
of a pretty-printed document. Previous versions of RapidJSON did not
flush when printing standalone top-level values, and the newline was
printed to the file, not the stream, so it could end up before the
values.

Fixes #4