Replace CMake with custom configure/make
CMake is gone (finally.) A handrolled configure script fetches
dependencies and supports --host and --prefix, the minimum necessary for
cross-compilation.

The app binaries are now each built completely in a single translation
unit instead of building each source file separately.

The generated man pages are now committed to the repository so they can
be installed from source without needing md2man. (md2man is still needed
to generate them when changes to the markdown are made.)